spike(console): question cards as a scrollable overlay inside the spec body
Move the collab question card from a full-width banner above the body
(which pushed the spec down) to a floating overlay anchored INSIDE the
body's content pane:

- body flex container gets position:relative; the overlay is absolutely
  positioned, inset past the 280px file list, bottom-anchored and centred.
- pointer-events are transparent except on the card, so the spec behind
  stays usable.
- sticky header + internal scroll region (maxHeight 60vh) so a long list
  of questions (ask_questions batch form) scrolls within the card.

SpecQuestionBanner.tsx → SpecQuestionOverlay.tsx. 386 tests pass;
typecheck + lint clean.
